+++ config/bootstrap.php
@@ -14,9 +14,11 @@ use Yiisoft\Config\Config;
 return [
     static function (ContainerInterface $container) use ($config, $params) {
         WidgetFactory::initialize(
-            $container,
-            $config->get($params['yiisoft/widget']['configGroup']),
-            $params['yiisoft/widget']['validateConfig']
+            container: $container,
+            definitions: $config->get($params['yiisoft/widget']['config']['definitionsGroup']),
+            validate: $params['yiisoft/widget']['config']['validate'],
+            themes: $config->get($params['yiisoft/widget']['config']['themesGroup']),
+            defaultTheme: $params['yiisoft/widget']['defaultTheme'],
         );
     },
 ];