+++ config/params.php
@@ -4,7 +4,11 @@ declare(strict_types=1);
 
 return [
     'yiisoft/widget' => [
-        'configGroup' => 'widgets',
-        'validateConfig' => true,
+        'config' => [
+            'definitionsGroup' => 'widgets',
+            'themesGroup' => 'widgets-themes',
+            'validate' => true,
+        ],
+        'defaultTheme' => null,
     ],
 ];